# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SourceStream API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get list of articles",
                "operationId": "93ded9e31119e6bbb22455c72ef97f19",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search by title, summary or content",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Game"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "search by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "business",
                                "entertainment",
                                "general",
                                "health",
                                "science",
                                "sports",
                                "technology"
                            ],
                            "example": "technology"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "search by source",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "newsapi,guardian,nytimes"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "published_at start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2025-09-26 07:28:50"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "published_at end date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2025-09-26 12:03:42"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Articles"
                    }
                }
            }
        },
        "/api/articles/{article}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get an Article by ID",
                "operationId": "be7af7750efdb2560ecfae862b170bf4",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article Data"
                    }
                }
            }
        },
        "/api/articles/filters/get": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get the filter option for Articles",
                "operationId": "6353492c117ba199e626f6a5eddf1203",
                "responses": {
                    "200": {
                        "description": "Articles Filter"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        }
    ]
}